add helpers to keep fetch clean
